refactor error to separate file
