+++ src/bls_ops.rs
@@ -1,11 +1,12 @@
 use crate::allocator::{Allocator, Atom, NodePtr};
 use crate::cost::{check_cost, Cost};
 use crate::err_utils::err;
+use crate::error::EvalErr;
 use crate::op_utils::{
     atom, first, get_args, get_varargs, int_atom, mod_group_order, new_atom_and_cost, nilp, rest,
     MALLOC_COST_PER_BYTE,
 };
-use crate::reduction::{EvalErr, Reduction, Response};
+use crate::reduction::{Reduction, Response};
 use chia_bls::{
     aggregate_pairing, aggregate_verify, hash_to_g1_with_dst, hash_to_g2_with_dst, G1Element,
     G2Element, PublicKey,
