+++ src/err_utils.rs
@@ -1,5 +1,5 @@
 use crate::allocator::NodePtr;
-use crate::reduction::EvalErr;
+use crate::error::EvalErr;
 
 pub fn err<T>(node: NodePtr, msg: &str) -> Result<T, EvalErr> {
     Err(EvalErr(node, msg.into()))
